vmap fallback: gracefully error out when vmap over dim of size 0
Previously, this would crash with a floating point error. If the user vmaps
over a dimension of size 0, ideally we would return a tensor with a
batch dim of size 0 and the correct output shape. However, this isn't
possible without a shape-checking API. This PR changes the vmap fallback
to error out gracefully if it sees vmap occuring over a dimension of
size 0.

If we want to support vmapping over dimension of size 0 for a specific
op, then the guidance is to implement a batching rule for that op that
handles 0-sized dims.

Test Plan:
- new test

ghstack-source-id: 4c84b66ed48cf5746574c79d9a99ad991a70e744
Pull Request resolved: https://github.com/pytorch/pytorch/pull/46846