[JIT] Add selective backend lowering API
**Summary**
This commit adds a helper/utility to faciliate the selective lowering of
specific submodules within a module hierarchy to a JIT backend. The reason
that this is needed is that lowering a submodule of a scripted
module to a backend after the module has been scripted requires
adjusting its JIT type.

**Test Plan**
This commit refactors `NestedModuleTest` in `jit/test_backends.py` to
use this new selective lowering API.

**Fixes**
This commit fixes ##41432.

ghstack-source-id: 57d0b51c4a71b8919fd1f29b9a70cfbbd56dbb8e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43613