[quant][graphmode][fx] Remove inplace option for convert_fx
Summary:
Initially we were thinking of adding a `invalidate_quantized_float_parameters` option to free the memory
of quantized floating parameters, but it turns out we will do module swap just like in eager mode for the modules
that are quantized, so the old floating point module will not be referenced after quantization. therefore this feature
is only needed for functionals, since most people are using quantization with modules we may not need this.

we'll revisit after we find there is a need for this.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 2c30a71933ac63ef4a7b46e0b0902a1652d9d3de
Pull Request resolved: https://github.com/pytorch/pytorch/pull/46955