[jit] introduce CallFunctionAsync to simplify fork compilation
Right now, we are creating an independent closure + fork subgraph for
every call to `torch.jit.fork()`. This balloons compile times
substantially when there are a lot of fork calls, since compiling
closures is pretty slow (as it involves creating a whole sbugraph).

This PR introduces an optimization for the non-closure case, where we
are just forking a function, like:

```
torch.jit.fork(my_func, arg1, arg2)
```

The above code will now generate:
```
%ret : Future[whatever] = prim::CallFunctionAsync(...)
```

instead of `prim::fork` and a closure subgraph to execute.

In practice, we only ever fork functions and methods, since closures are
not possibly to create in the Python frontend. Existing serialized code
will still contain closures as that is how we serialized forked code
previously.

ghstack-source-id: 6fada3dadd14691f65fae3bed807fc24af4a7ff3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47078